changed userId to type long and added endpoint for highscore update plus userService class

+++ backend/src/main/java/com/example/journey/autumn/security/UserPrincipal.java
@@ -51,5 +51,5 @@ public class UserPrincipal implements UserDetails {
     public boolean isEnabled() {
         return true;
     }
-    public int getId() { return user.getId(); }
+    public long getId() { return user.getId(); }
 }
